Add files via upload
This is the main code, incorporating the environment, the necessary ply files we want, and the gripper working. Currently functioning:
- Working gripper
- Cannonball and Powder ball are moving with the vertices commands
Some things I reckon we will need -> Collision detection and path finding, as the arm makes contact to the cannon easily, might need to resize the cannon also. 
Discuss with me in class if you need help with anything